Create a subpackage for containerized undercloud dependencies
We had to rework this because initially we did it in instack-undercloud
but Dan Prince highlighted the fact that the project aim to disappear
one day so let's start by not adding new things.

This patch adds the dependencies of a containerized undercloud into a
new subpackage: python-tripleoclient-heat-installer.

For now, this is an optional subpackage but in Rocky we'll deploy the
dependencies by default for backward compatibility.

Co-authored-by: Bogdan Dobrelia <bdobreli@redhat.com>
Change-Id: I97ec756ad116e1c86c7851f259a36be4a118bd11