Create a subpackage for containerized undercloud dependencies
We had to rework this because initially we did it in instack-undercloud
but Dan Prince highlighted the fact that the project aim to disappear
one day so let's start by not adding new things.

This patch adds the dependencies of a containerized undercloud into a
new subpackage: python-tripleoclient-heat-installer and require this
dependency to be deployed when deploying tripleoclient.

That way, the deployment of containerized undercloud should not add new
packages that operators need to install, the dependencies will take care
of it.

Also since it's a subpackage, it's pretty easy to ignore it if really
needed by operators.

Change-Id: I97ec756ad116e1c86c7851f259a36be4a118bd11