Added new requirement for tripleoclient.
Adding this requirement for the tripleoclient will allow to get
a proper preflight validation. The tripleo-validations has many
interesting validation ansible playbooks, and we want to use
some of them, like (not complete):
- disk space on undercloud
- ram available on the undercloud

Currently, this "tripleo-validations" package is added only during
the undercloud deploy, but we need it prior, in order to use its
content.

Change-Id: I9b4c184096c72fb5367e65e103df15096d2f6cb3