Add xfsprogs to the required packages
The recent purge of docker from the system deprived the host system from
and important dependency that was pulled during its installation.
xfsprogs is not installed anymore and it's not an essential package in
the base cloud image (for more or less obvious reasons)
This patch adds xfsprogs to the requirement for tripleoclient, which is
for now the safest place where to put it. Diskimage builder is a
multiplatform tool, and other filesystems may be used instead of xfs, so
it could not be "required" then, only optional

Change-Id: I516dd7b26aecb7c658d9ff11bb17e78c87cdbe3d