Update patch set 10
Patch Set 10: Code-Review+1

The exit 0 caused multiple lines to short-circuit the first time a check succeeded.  This edit deletes the exit 0 at the end of the macro which enables the %pre (or %pretrans) script to continue rather than terminating early the moment the correct module is found.

Patch-set: 10
Label: Code-Review=+1