Update patch set 13
Patch Set 13:

The "add help info" was based on IRC chatter.

Throwing an error after RPM dependencies are resolved but during the RPM transaction (after some dependencies may have been installed) is kind of ugly, but there's no other clean way to do it since RPMs cannot depend directly on modules.

So, the thought is, since we're doing something that's somewhat hostile, that we should at least tell people how to fix the problem.

As an aside, the intent here is to hide this macro in some other build macro package (openstack-macros? RPM base?) but only after this is agreed to be working and what we want.

Patch-set: 13
Reviewer: Gerrit User 363 <363@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0