Incrementing the validations-libs version to 1.2.0
The valdiations-libs 1.2.0 introduces several changes
scheduled for the 16.2 and 17.0 releases.

The previously merged, and reverted, patch sadly indicated version 1.2.1.
Which was incorrect and could have caused issues down the road.

Nevertheless the version bump is necessary to ensure things are up to date,
as the validations-libs release was recently defined in the relevant distgit.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I417305dadc15a0f3683ceaabdc611790d928d84b
(cherry picked from commit 961636fb2030504b78549a2a8e1932f987405f5b)