Remove redhat-lsb-core as build requirement
The package redhat-lsb-core has been retired from RHEL9. This patch is
removing it as BR as i guess this is just for historical reasons as it's
mainly a metapackage used to install basic tools related to lsb
certification. If any of those utilities are required to build the
package, they should be included as BRs instead of redhat-lsb-core.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2012924

Change-Id: Idce26bb9fa93f7d9b9e86f2fb8f45331c9d1091f